Reworked script to reconnect only buggy bluetooth devices
Reconnection of all trusted devices is not so good idea generally, because
paired devices should reconnect automatically.
However there are exists multiple devices which for some reason tries to,
but can not reconnect properly. From the host perspective it looks like
immediate disconnect.
With this change we trying to catch such immediate disconnects and force
reconnect such devices.

Signed-off-by: Yuri Konotopov <ykonotopov@gnome.org>